Added calendar
Added calendar and most of the basic calendar functionalities now also update the back end. Only need to configure the layout a bit more to give more space to the visual side of the bookings and make the dragging out of bookings actually update towards the back end. Drag and drop for bookings is already functional.